Precursor for password encoding. Not yet working.

+++ build.ps1
@@ -9,7 +9,9 @@ New-Item -ItemType Directory -Path ".\build"
 Set-Location ".\build"
 cmake -G "MinGW Makefiles" ..
 cmake --build . --config Release
-Move-Item ".\codify.exe" "..\codify.exe" -Force
+if (Test-Path ".\codify.exe") {
+    Move-Item ".\codify.exe" "..\codify.exe" -Force
+}
 if (Test-Path ".\codify") {
     Move-Item ".\codify" "..\codify" -Force
 }
